+++ package.json
@@ -18,7 +18,7 @@
     "@nuxt/fonts": "^0.14.0",
     "@nuxtjs/color-mode": "^4.0.0",
     "@unocss/preset-icons": "^66.6.0",
-    "@unveil/identity": "1.12.0-beta.0",
+    "@unveil/identity": "1.12.0-beta.1",
     "@vueuse/core": "^14.2.1",
     "dayjs": "^1.11.19",
     "nuxt": "^4.4.6",
